requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
PyMuPDF==1.23.8
python-engineio==4.8.0
python-socketio==5.10.0
eventlet==0.33.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-CORS==4.0.0